split logic ans styles of profile page
