+++ src/pages/OfferModal/ProfilePage/ProfilePage.jsx
@@ -1,19 +1,10 @@
-import React, { useContext, useEffect, useState } from 'react';
+import React from 'react';
 
 import styles from './ProfilePage.module.scss';
 
-import ProfileItem from './ProfileItem';
-import useForm from './SubPages/useForm';
-import FORM from './form';
-import createBirthdayLabel from './utils/createBirthdayLabel';
-import LocalStorage from './utils/LocalStorage';
 import Button from '../../../components/Button';
-import { AuthContext } from '../../../auth/Auth';
-import { saveProfile } from '../../../apis';
 
-const SAVE_PROFILE_FAIL = 'Profile saving failed. Please try again later.';
-
-const { getStoredData, storeData, dropStoredData } = new LocalStorage('userProfile');
+const ProfilePage = {};
 
 const Header = ({ onGoBack }) => (
   <>
@@ -49,108 +40,8 @@ const Footer = ({
   </Button>
 );
 
-export default function useProfilePage(
-  showMessage, setProfileExist,
-) {
-  const { currentUser } = useContext(AuthContext);
-  const userId = currentUser && currentUser.userId;
+ProfilePage.Header = Header;
+ProfilePage.Content = Content;
+ProfilePage.Footer = Footer;
 
-  const [subPage, loadSubPage] = useState();
-
-  const [profileFilled, setProfileFilled] = useState(false);
-
-  const form = useForm(FORM, getStoredData());
-
-  const {
-    getData,
-    handleDataChange,
-    findEmptyField,
-  } = form;
-
-  const formData = getData();
-
-  const {
-    photo,
-    bankAccount,
-    billingAddress,
-    birthday,
-    mobile,
-  } = formData;
-
-  const checkProfile = () => {
-    if (!findEmptyField()) setProfileFilled(true);
-  };
-
-  useEffect(() => {
-    storeData(formData);
-    checkProfile();
-  }, [
-    photo,
-    bankAccount,
-    billingAddress,
-    birthday,
-    mobile,
-  ]);
-
-  const goBack = () => {
-    loadSubPage('');
-  };
-
-  const profileList = (
-    <div className={styles.profile_list_wrapper} >
-      {Object.keys(FORM).map((key) => {
-        const { label, Page } = FORM[key];
-        const value = formData[key];
-
-        const handleChange = (input) => {
-          handleDataChange(key)(input);
-          goBack();
-        };
-
-        const statusLabel = {
-          birthday: createBirthdayLabel(value),
-          mobile: value,
-        }[key] || null;
-
-        return (
-          <ProfileItem
-            itemName={label}
-            handleClick={() => loadSubPage(
-              <Page
-                storedValue={value}
-                onSubmit={handleChange}
-              />
-            )}
-            statusLabel={statusLabel}
-            filled={!!value}
-            key={key}
-          />
-        );
-      })}
-    </div>
-  );
-
-  const submitProfile = async () => {
-    const result = await saveProfile(userId, formData);
-
-    if (result) {
-      setProfileExist(true);
-      // dropStoredData();
-    } else {
-      showMessage(SAVE_PROFILE_FAIL);
-    }
-  };
-
-  const page = {};
-  page.header = <Header goBack={subPage && goBack} />;
-  page.content = <Content content={subPage || profileList} />;
-  page.footer = (
-    <Footer
-      onBottomClick={subPage ? goBack : submitProfile}
-      isBottomBtnDisabled={!subPage && !profileFilled}
-      btnLabel={subPage ? 'Back' : 'Continue'}
-    />
-  );
-
-  return page;
-}
+export default ProfilePage;
