+++ src/pages/OfferModal/ProfilePage/index.js
@@ -1,1 +1,1 @@
-export { default } from './ProfilePage';
+export { default } from './useProfilePage';
